requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
cryptography
flask==3.0.2
Flask-MySQL==1.5.2
flask-mysql-connector==1.1.0
Flask-RESTful==0.3.10
flask-sqlalchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
importlib-metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.24.4
pandas==2.0.3
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.28
typing-extensions==4.10.0
tzdata==2024.1
waitress
werkzeug==3.0.1
zipp==3.18.1
